Rollback individual transaction changes
