+++ data-tx/src/main/java/io/micronaut/transaction/test/DefaultTestTransactionExecutionListener.java
@@ -20,6 +20,7 @@ import io.micronaut.context.annotation.Property;
 import io.micronaut.context.annotation.Replaces;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.core.annotation.Internal;
+import io.micronaut.core.annotation.Nullable;
 import io.micronaut.core.util.StringUtils;
 import io.micronaut.spring.tx.test.SpringTransactionTestExecutionListener;
 import io.micronaut.test.annotation.TransactionMode;
@@ -27,7 +28,13 @@ import io.micronaut.test.context.TestContext;
 import io.micronaut.test.context.TestExecutionListener;
 import io.micronaut.test.extensions.AbstractMicronautExtension;
 import io.micronaut.transaction.SynchronousTransactionManager;
-import jakarta.inject.Provider;
+import io.micronaut.transaction.TransactionDefinition;
+import io.micronaut.transaction.TransactionStatus;
+import io.micronaut.transaction.sync.SynchronousFromReactiveTransactionManager;
+
+import java.lang.reflect.AnnotatedElement;
+import java.lang.reflect.Method;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Adds support for MicronautTest transactional handling.
@@ -41,58 +48,94 @@ import jakarta.inject.Provider;
 @Replaces(SpringTransactionTestExecutionListener.class)
 @Internal
 public class DefaultTestTransactionExecutionListener implements TestExecutionListener {
-    private final TestExecutionListener delegate;
+    private final SynchronousTransactionManager<Object> transactionManager;
+    private final TransactionMode transactionMode;
+    private TransactionStatus<Object> tx;
+    private final AtomicInteger counter = new AtomicInteger();
+    private final AtomicInteger setupCounter = new AtomicInteger();
+    private final boolean rollback;
+    @Nullable
+    private final SpockMethodTransactionDefinitionProvider spockMethodTransactionDefinitionProvider;
 
     /**
-     * @param transactionManagerProvider transaction manager provider
-     * @param rollback                   {@code true} if the transaction should be rollback
-     * @param transactionMode            The transaction mode
+     * @param transactionManager                       Spring's {@code PlatformTransactionManager}
+     * @param rollback                                 {@code true} if the transaction should be rollback
+     * @param transactionMode                          The transaction mode
+     * @param spockMethodTransactionDefinitionProvider The spock method name extractor
      */
-    DefaultTestTransactionExecutionListener(
-            Provider<SynchronousTransactionManager<Object>> transactionManagerProvider,
-            @Property(name = AbstractMicronautExtension.TEST_ROLLBACK, defaultValue = "true") boolean rollback,
-            @Property(name = AbstractMicronautExtension.TEST_TRANSACTION_MODE, defaultValue = "SEPARATE_TRANSACTIONS") TransactionMode transactionMode) {
+    protected DefaultTestTransactionExecutionListener(
+        SynchronousTransactionManager<Object> transactionManager,
+        @Property(name = AbstractMicronautExtension.TEST_ROLLBACK, defaultValue = "true") boolean rollback,
+        @Property(name = AbstractMicronautExtension.TEST_TRANSACTION_MODE, defaultValue = "SEPARATE_TRANSACTIONS") TransactionMode transactionMode,
+        @Nullable SpockMethodTransactionDefinitionProvider spockMethodTransactionDefinitionProvider) {
+        this.spockMethodTransactionDefinitionProvider = spockMethodTransactionDefinitionProvider;
 
-        if (transactionMode == TransactionMode.SINGLE_TRANSACTION) {
-            SynchronousTransactionManager<Object> transactionManager = transactionManagerProvider.get();
-            delegate = new DefaultTestSingleTransactionExecutionListener(transactionManager, rollback);
+        if (transactionManager instanceof SynchronousFromReactiveTransactionManager) {
+            throw new IllegalStateException("Transaction mode is not supported when the synchronous transaction manager is created using Reactive transaction manager!");
+        }
+
+        this.transactionManager = transactionManager;
+        this.rollback = rollback;
+        this.transactionMode = transactionMode;
+    }
+
+    @Override
+    public void beforeSetupTest(TestContext testContext) {
+        beforeTestExecution(testContext);
+    }
+
+    @Override
+    public void afterSetupTest(TestContext testContext) {
+        if (transactionMode.equals(TransactionMode.SINGLE_TRANSACTION)) {
+            setupCounter.getAndIncrement();
         } else {
-            // TestMethodInterceptor is handing transactions
-            delegate = new NoopTestExecutionListener();
+            afterTestExecution(false);
         }
     }
 
     @Override
-    public void beforeSetupTest(TestContext testContext) throws Exception {
-        delegate.beforeSetupTest(testContext);
+    public void beforeCleanupTest(TestContext testContext) {
+        beforeTestExecution(testContext);
     }
 
     @Override
-    public void afterSetupTest(TestContext testContext) throws Exception {
-        delegate.afterSetupTest(testContext);
+    public void afterCleanupTest(TestContext testContext) {
+        afterTestExecution(false);
     }
 
     @Override
-    public void beforeCleanupTest(TestContext testContext) throws Exception {
-        delegate.beforeCleanupTest(testContext);
+    public void afterTestExecution(TestContext testContext) {
+        if (transactionMode.equals(TransactionMode.SINGLE_TRANSACTION)) {
+            counter.addAndGet(-setupCounter.getAndSet(0));
+        }
+        afterTestExecution(this.rollback);
     }
 
     @Override
-    public void afterCleanupTest(TestContext testContext) throws Exception {
-        delegate.afterCleanupTest(testContext);
+    public void beforeTestExecution(TestContext testContext) {
+        if (counter.getAndIncrement() == 0) {
+            TransactionDefinition definition;
+            AnnotatedElement annotatedElement = testContext.getTestMethod();
+            if (spockMethodTransactionDefinitionProvider != null) {
+                definition = spockMethodTransactionDefinitionProvider.provide(annotatedElement);
+            } else if (annotatedElement instanceof Method) {
+                Method method = (Method) annotatedElement;
+                String name = method.getDeclaringClass().getSimpleName() + "." + method.getName();
+                definition = TransactionDefinition.named(name);
+            } else {
+                definition = TransactionDefinition.DEFAULT;
+            }
+            tx = transactionManager.getTransaction(definition);
+        }
     }
 
-    @Override
-    public void beforeTestExecution(TestContext testContext) throws Exception {
-        delegate.beforeTestExecution(testContext);
+    private void afterTestExecution(boolean rollback) {
+        if (counter.decrementAndGet() == 0) {
+            if (rollback) {
+                transactionManager.rollback(tx);
+            } else {
+                transactionManager.commit(tx);
+            }
+        }
     }
-
-    @Override
-    public void afterTestExecution(TestContext testContext) throws Exception {
-        delegate.afterTestExecution(testContext);
-    }
-
-    private static final class NoopTestExecutionListener implements TestExecutionListener {
-    }
-
 }
