+++ doc-examples/jdbc-and-r2dbc-example-java/src/test/java/example/BookJdbcRepositorySpec.java
@@ -14,7 +14,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-@MicronautTest
+@MicronautTest(transactional = false)
 class BookJdbcRepositorySpec {
 
 	@Inject BookJdbcRepository bookRepository;
