+++ doc-examples/jdbc-example-kotlin/src/test/kotlin/example/ProjectConfig.kt
@@ -1,0 +1,9 @@
+package example
+
+import io.kotest.core.config.AbstractProjectConfig
+import io.micronaut.test.extensions.kotest.MicronautKotestExtension
+
+object ProjectConfig : AbstractProjectConfig() {
+    override fun listeners() = listOf(MicronautKotestExtension)
+    override fun extensions() = listOf(MicronautKotestExtension)
+}
